Move connectors to DB folder
